added a function that should remove small floating point inaccuracies
added a function that checks the results from the manipulated transform matrix.
the function checks if the new values have a meaningful difference from the original.
this function was added as the scale seemed to decrease by 0.00001 every frame which could eventually cause a noticeable differnce in the scale overtime