mouse picking has been implemented
When hovering over an entity the tag of the entity will displayed in the stats window.
The entity ID is save into the framebuffer attachment as an unsigned integer that can then be read from the attachment using the readPixel function.
Added Unsigned int version of a various macros and buffer types so that unsinged ints can be sent to and read from the gpu textures.
Added various TO DO comments that can be followed to expanded the API and make it more usable.