started to implement new shader system using vulkan and spir-v
added setup script - requires python 3, if python is not installed when using setup.bat the user will be told and linked to the python download website.
The setup script will check if the vulkan SDK is present on the machine and if its not then it will run the installation for it (version 1.4.321.1)
The vulkan sdk is required for converting shaders into vulkan -> spir-v -> openGL
This does mean that the openGL shaders are dependant on vulkan so i might want to cosider later compiling spir-v on it's own into the project.
Still need to implement uniform buffers before this is functional