expanded content browser panel, better layout, more functional
The content browser now has directory and file icons. add the ability to drag and drop a scene file into the viewport and it will open the scene.
Separated files: Resources are used by the engine, Assets are used from a game/project
added a try statement to deserialize so if anything goes wrong e.g. a .png is passed in then it will fail but the program an carry on.