Circle rending has been added
The is a bug when compiling the circle shader, when using discard spirv_cross throws a memory access violation. If the program is just continued the shader functions as expected.
Investigate the reason discard is causing the compilation error or find an alternative to fix the issue

removed separators between components (Inspector panel)